# Speed up nonce service.