--- a/nonce/nonce_test.go
+++ b/nonce/nonce_test.go
@@ -88,3 +88,33 @@ func TestRejectTooEarly(t *testing.T) {
 	test.Assert(t, ns.Valid(n1), "Rejected a valid nonce")
 	test.Assert(t, !ns.Valid(n0), "Accepted a nonce that we should have forgotten")
 }
+
+func BenchmarkNonces(b *testing.B) {
+	ns, err := NewNonceService(metrics.NewNoopScope())
+	if err != nil {
+		b.Fatal("creating nonce service", err)
+	}
+
+	for i := 0; i < ns.maxUsed; i++ {
+		n, err := ns.Nonce()
+		if err != nil {
+			b.Fatal("noncing", err)
+		}
+		if !ns.Valid(n) {
+			b.Fatal("generated invalid nonce")
+		}
+	}
+
+	b.ResetTimer()
+	b.RunParallel(func(pb *testing.PB) {
+		for pb.Next() {
+			n, err := ns.Nonce()
+			if err != nil {
+				b.Fatal("noncing", err)
+			}
+			if !ns.Valid(n) {
+				b.Fatal("generated invalid nonce")
+			}
+		}
+	})
+}