@rolandshoemaker 2018-03-12
Did you write a golang benchmark for this? If so could you check it in (and it'd be great to see the [`benchcmp`](https://godoc.org/golang.org/x/tools/cmd/benchcmp) output in the PR description).