@jsha 2018-03-12
> Did you write a golang benchmark for this?

Unfortunately, no. I wrote a standalone main.go when trying to reproduce the main problem, but got hung up trying to reproduce the blocking issues rather than measure the raw speed of the operation. I'll write a quickie benchmark for this now.